Extreme Alpha Build - Working Basics Out
Code is still sloppy, next update will clean up code. This version is
working, however, and needed to be backed up before more work. This
version includes 1) A mini 'Town' area where the player can either move
to edit the level or play the level. Editing the level will allow a top
down birds eye view where maze pieces can be placed. Playing the level
will allow the player to run through the maze. 2) Maze can be built
using many pieces. Pieces can be deleted and replaced. 3) End technology
calculates where dead ends need to be as opposed to open hallways 4)
Maps can be saved with P and loaded with L in edit mode, saved map is
the one that will be played when going to play the level. Probably more
to this version, but all I can think of right now.